@@ -8,8 +8,9 @@
     }
   ],
   "require": {
-    "php": "^5.6|^7.0|^8.0",
-    "ext-json": "*"
+    "php": "^7.4|^8.0",
+    "ext-json": "*",
+    "ext-dom": "*"
   },
   "require-dev": {
     "phpunit/phpunit": "^5.6|^6.0|^7.0|^8.0|^9.0"
